+++ src/pages/data/data.css
@@ -1,0 +1,3 @@
+.video {
+    text-align: center;
+}
