+++ src/pages/index.js
@@ -35,76 +35,13 @@ const IndexPage = () => (
           </div>
           <h1>What do you want to find out?</h1>
           <p>
-            It is well established that women are under-represented in
-            philosophy at all levels. There is also some data on the
-            under-representation of women in journals. However, further data is
-            necessary to see more clearly what has gone wrong, how things have
-            been changing over time, and to understand how women are represented
-            within different journals, professorial types, and ranks at
-            different kinds of departments.{" "}
+          For the past decade or so, the lack of diversity in philosophy has received considerable attention. Though a number of studies have shown that there has been some progress made in the demographics of philosophy in the past decade, the progress is slow and disparities persist. (See for example, <a href="https://www.journals.uchicago.edu/doi/full/10.1086/718075">Hassoun, Conklin, Nekrasov, and West 2022</a>; <a href="https://www.philosophersmag.com/essays/244-the-diversity-of-philosophy-students-and-faculty-in-the-united-states">Schwitzgebel, Bright, Jennings, Thompson, and Winsberg 2021</a>; and other <Link to ="resources">resources on the resources page</Link>.)
           </p>
           <p>
-            This site provides some new <Link to="data">data</Link> on disparities
-            between the proportion of women and men in faculty hiring and
-            publishing.
+            A culture change is needed in philosophy. To foster this change we, the Demographics in Philosophy project, have done a few things. First, we have collected existing data on the representation of women in philosophy and put together a list of data based resources that may help us see more clearly what has gone wrong, how things have been changing over time, and to understand representation within different journals, professorial types, and ranks at different kinds of departments. On this site you can explore data on disparities between the proportion of those who are, and those who are not, under-represented in faculty hiring and publishing. You find resources for increasing diversity in philosophy and bibliographical information on surveys and large-scale empirical studies on the problem of underrepresentation in philosophy.
           </p>
           <p>
-            We also provide some further <Link to="resources">Resources</Link> for
-            those who want to make a difference for women in philosophy. For
-            some summary information, see{" "}
-            <a href="https://blog.apaonline.org/2018/03/26/tell-us-how-to-fix-the-lack-of-diversity-in-philosophy-journals/">
-              this blog post
-            </a>
-            ,{" "}
-            <a href="https://phys.org/news/2017-05-women-underrepresented-philosophy-journals-reveals.html">
-              this headline
-            </a>
-            ,{" "}
-            <a href="https://link.springer.com/article/10.1007/s11098-017-0919-0">
-              this article
-            </a>
-            ,{" "}
-            <a href="http://scienmag.com/new-website-uses-big-data-to-address-underrepresentation-of-women-in-philosophy/">
-              the news
-            </a>
-            ,{" "}
-            <a href="http://blog.apaonline.org/2016/11/15/can-data-help-us-improve-the-situation-for-women-in-philosophy/">
-              this post
-            </a>
-            ,
-            <a href="http://blog.apaonline.org/2017/05/25/how-to-improve-the-situation-for-women-in-philosophy/">
-              {" "}
-              this
-            </a>
-            ,{" "}
-            <a href="https://blog.apaonline.org/2018/08/23/diversity-and-philosophy-journals-introduction/">
-              this one
-            </a>
-            ,{" "}
-            <a href="https://blog.apaonline.org/2018/08/30/diversity-and-philosophy-journals-how-to-avoid-conservative-gatekeeping/">
-              this journal
-            </a>
-            ,{" "}
-            <a href="https://blog.apaonline.org/2018/09/06/diversity-and-philosophy-journals-sophias-history-of-encouraging-diversity/">
-              this contribution
-            </a>
-            ,{" "}
-            <a href="https://blog.apaonline.org/2018/09/13/diversity-and-philosophy-journals-some-comments-on-diversity/">
-              this comment
-            </a>
-            ,{" "}
-            <a href="https://blog.apaonline.org/2018/09/20/diversity-and-philosophy-journals-reforming-philosophy/">
-              this reform
-            </a>
-            ,{" "}
-            <a href="https://blog.apaonline.org/2018/10/01/diversity-and-philosophy-journals-diverse-aims-at-mind/">
-              this review
-            </a>
-            , and{" "}
-            <a href="https://blog.apaonline.org/2018/10/04/diversity-and-philosophy-journals-practices-for-improving-diversity-in-philosophy-journal-publishing/">
-              this practice
-            </a>
-            .
+            We have also held a series of listening sessions with journals and department heads on best practices and reviewed the literature to create a set of Good Practice Guidelines that departments, journals, and societies can <Link to="practices">sign up to</Link>– also available as a <Link to="TODOFIXME">downloadable document</Link>. The aim of this guide is to address factors that affect inclusion of underrepresented groups in philosophy - such as the methods of practicing philosophy, how it is taught, the content of philosophy and the workings of philosophy programs and their practices. The Good Practice Guidelines draw, and expand, on the APA’s Good Practices Guide and Blog on Diversity and Equity in Recruitment, the BPA/SWIP Good Practice Scheme, the Oxford University faculty of law Good Practices Guide, and the Canadian Society for Women in Philosophy Guidelines for Conference Hosting.
           </p>
